+++ test/unit/test_cmd_engine.py
@@ -5,4 +5,4 @@
 
 import pytest
 
-from mini_project_1.cmd_engine import ProjectShell
+from mini_project_1.project_shell import ProjectShell
